requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
deepface==0.0.93
filelock==3.16.1
fire==0.7.0
flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
gast==0.4.0
gdown==5.2.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.1
gunicorn==23.0.0
h5py==3.11.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
mtcnn==0.1.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.4.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1-modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soupsieve==2.6
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing-extensions==4.5.0
tzdata==2024.2
urllib3==2.2.3
werkzeug==3.0.6
wrapt==1.17.0
zipp==3.20.2
